// CaffStore.Frontend/src/app/api/generated/model/commentDto.ts
/**
 * Caff Store API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDto } from './userDto';


export interface CommentDto { 
    id: number;
    text: string;
    createdAt: string;
    createdBy: UserDto;
    lastModifiedAt: string;
}